protein-loc-scaling/requirements.txt
# core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3

# deep learning (for embedding extraction)
torch>=2.0
transformers>=4.30

# additional classifiers
xgboost>=2.0

# visualization
matplotlib>=3.7
seaborn>=0.12

# dimensionality reduction
umap-learn>=0.5

# statistics
scipy>=1.10
